syslog: truncate the length argument to the C int Linux declares
`SYSCALL_DEFINE3(syslog, int type, char __user *, buf, int len)` takes the
length as an `int`, so Linux truncates the register to its low 32 bits before
do_syslog()'s `if (!buf || len < 0) return -EINVAL` (kernel/printk/printk.c).
The kernel declared the parameter as `isize` and passed the register through
untruncated.

The x86-64 psABI lets a variadic `int` argument arrive zero-extended, so a
caller writing `syscall(SYS_syslog, SYSLOG_ACTION_READ_ALL, buf, -1)` -- the
form the guest differential uses -- presents 0x00000000ffffffff.  Read as a
64-bit signed length that is a huge positive count rather than -1, so the
negative-length test was skipped and the request reached the accessibility
check, answering EFAULT where Linux answers EINVAL.  This is what stopped the
fs-abi differential at syslog.raw-differential negative-len.

The policy crate was already correct; only the argument decoding was wrong,
which is why the crate's own unit tests passed throughout.

+++ kernel/src/syscall/dispatch.rs
@@ -1486,12 +1486,19 @@ pub(super) fn dispatch_syscall(
         Sysno::sysinfo => {
             with_user_memory(aspace(), |memory| sys_sysinfo(memory, uctx.arg0() as _))
         }
+        // `SYSCALL_DEFINE3(syslog, int type, char __user *, buf, int len)` takes
+        // the length as a C `int`, so the register is truncated to its low 32
+        // bits before `do_syslog()`'s `len < 0` test.  Without the truncation a
+        // caller that passes a negative `int` -- which the x86-64 ABI lets
+        // arrive zero-extended -- is read as a huge positive length, so it
+        // reaches the accessibility check and yields EFAULT where Linux's
+        // negative-length test yields EINVAL.
         Sysno::syslog => with_user_memory(aspace(), |memory| {
             sys_syslog(
                 memory,
                 uctx.arg0() as _,
                 uctx.arg1() as _,
-                uctx.arg2() as isize,
+                uctx.arg2() as u32 as i32 as isize,
             )
         }),
         Sysno::getrandom => with_user_memory(aspace(), |memory| {
